fix: split TestLoadCoverageMap out of test_test_filter.py to stay under 1000-line limit
test_test_filter.py grew to 1002 lines after adding _HOOKS_UNCONDITIONAL_FILES guards.
Extracted TestLoadCoverageMap into test_test_filter_coverage_map.py (natural boundary
matching the existing split pattern) to bring test_test_filter.py to 931 lines.
